move common methods out of shed_install
